feat: add hook tests
